refactor(cbor_serialization): split up code by usecase and function
Properly split up the code by its usecase, resulting in "common",
"identity" and "record", as well as by whether the code is used for
serialization or deserialization. This makes the code way easier to
parse.
Also document this split by usecase in the README.
