+++ nodes/firmware/custom-modules/cbor_serialization/Makefile
@@ -1,3 +1,6 @@
 MODULE := cbor_serialization
 
+SRC += $(wildcard $(CURDIR)/serialization/*.c)
+SRC += $(wildcard $(CURDIR)/deserialization/*.c)
+
 include $(RIOTBASE)/Makefile.base
