+++ nodes/firmware/custom-modules/cbor_serialization/README.md
@@ -4,6 +4,12 @@
 
 The `cbor_serialization` module handles [CBOR](https://datatracker.ietf.org/doc/html/rfc8949) encoding and decoding of messages received via BLE (via [`mate_ble`](../mate_ble/)) and LoraWAN (via [`mate_lorawan`](../mate_lorawan/)). The `identity_store` module also heavily utilizes it for CBOR encoding and decoding.
 
+The code is split in the following logical parts:
+
+- `common`: Common functionality relied upon by various other parts of this module. (Like version and message type definitions as well as functions to retrieve those.)
+- `identity`: Functions to serialize and deserialize identities, identity messages as well as provisioning date. Used by the gossiping of identities (in `mate_ble`) as well as the `identity_store`.
+- `record`: Functions to serialize and deserialize records. Used by the record processing in `mate_ble` and `mate_lorawan`.
+
 ## Message Structure
 
 On the top level each message is a flat array of data with the version and message type information shared across all types of messages.
