+++ nodes/firmware/custom-modules/mate_lorawan/mate_lorawan.c
@@ -43,7 +43,8 @@
 #include "tables.h"
 #include "tables/records.h"
 #include "tables/types.h"
-#include "cbor_serialization.h"
+#include "cbor_serialization/record.h"
+#include "cbor_serialization/identity.h"
 #include "personalization.h"
 #include "event.h"
 #include "event/thread.h"
